fix: port wiki related functions to new file
